PROJ-17 Refactored the roles for QA MD deployment.
We have the role qa-md-appserver, which is generic across any QA deployment of Map Display.
And we have the role qa-md01-deployment, which is specific to the given deployment.

So there is a single qa-md-appserver role, which can be used for multiple QA deployments of Map Display, each of which has a role qa-mdXX-deployment, which defines things specific to that deployment, such as the database connection details and vhost.

git-svn-id: http://svn.mapofmedicine.com/map-sys/branches/CHEF_EXPERIMENTAL/chef-repo@2143 daf2e314-7611-4b85-95e2-b5f857fec4d4